chore: refactor workout to as a page
