+++ src/pages/workout/activeWorkout/exerciseSetEdit/ActiveWorkoutExerciseSetup.tsx
@@ -1,8 +1,8 @@
 import { useContext } from "react";
 import { ExerciseSetup } from "../workoutTypes";
 import { ActiveWorkoutSetContext } from "../ActiveWorkoutContext";
-import { ExerciseAutocomplete } from "../../../exerciseAutocomplete/ExerciseAutocomplete";
 import { RemoveRowDeleteIcon } from "../utils/RemoveRowDeleteIcon";
+import { ExerciseAutocomplete } from "../../../exercise/exerciseAutocomplete/ExerciseAutocomplete";
 
 const SetupItem = ({
   value,
