+++ src/pages/workout/activeWorkout/workoutTypes.ts
@@ -1,4 +1,4 @@
-import { MuscularGroupEnum } from "../../../../../utils/useMuscularGroupEnum";
+import { MuscularGroupEnum } from "../../../utils/useMuscularGroupEnum";
 
 export type ExerciseWithMuscularGroup = Exercise & {
   muscularGroup: MuscularGroupEnum;
