+++ test-setup.ts
@@ -1,6 +1,8 @@
 require("@testing-library/jest-dom");
 require("@testing-library/jest-dom/jest-globals");
 
+console.warn = jest.fn();
+
 window.matchMedia = jest.fn().mockReturnValue({
   matches: false,
   addListener: jest.fn(),
